paging without Paging3 lib

+++ app/src/main/java/com/example/allmyfriends/repository/PeopleRepository.kt
@@ -1,11 +1,11 @@
 package com.example.allmyfriends.repository
 
-import androidx.paging.*
 import androidx.room.withTransaction
 import com.example.allmyfriends.data.local.AllMyFriendsDatabase
 import com.example.allmyfriends.data.local.UserDao
 import com.example.allmyfriends.data.remote.ApiService
 import com.example.allmyfriends.model.User
+import com.example.allmyfriends.model.dto.Info
 import com.example.allmyfriends.util.Result
 import com.example.allmyfriends.util.networkBoundResource
 import kotlinx.coroutines.flow.Flow
@@ -13,30 +13,23 @@ import kotlinx.coroutines.flow.Flow
 class PeopleRepository(private var service: ApiService, var db: AllMyFriendsDatabase) {
     private var userDao: UserDao = db.personDao()
 
-    fun getUsers(): Flow<Result<PagingData<User>>> = networkBoundResource(
-        query = {
-            val pagingSourceFactory = { userDao.getPeople() }
-            Pager(
-                config = PagingConfig(pageSize = PAGE_SIZE, enablePlaceholders = true),
-                pagingSourceFactory = pagingSourceFactory
-            ).flow
-        },
+    fun getUsers(page: Int): Flow<Result<List<User>>> = networkBoundResource(
+        query = { userDao.getUsers() },
         shouldFetch = {
             //if (internetAvailable())
             true
         },
         fetch = {
-            service.queryData().users.map { it.toDomainModel() }
+            val apiResponse = service.queryData(page)
+            val info = apiResponse.info
+            apiResponse.users.map { it.toDomainModel(info.page) }
         },
         saveFetchResult = { people ->
             db.withTransaction {
-                userDao.deleteAllUsers()
-                userDao.insertPeople(people)
+                if (page == 0)
+                    userDao.deleteAllUsers()
+                userDao.insertUsers(people)
             }
         }
     )
-
-    companion object {
-        private const val PAGE_SIZE = 20
-    }
 }
